update video_utils.py and image_utils.py
